+++ src/onegov.core/onegov/core/framework.py
@@ -421,6 +421,13 @@ class Framework(TransactionApp, WebassetsApp, ServerApplication):
     @cached_property
     def translations(self):
         """ Returns all available translations keyed by langauge. """
+
+        if not self.registry.settings.i18n.domain:
+            return {}
+
+        if not self.registry.settings.i18n.localedir:
+            return {}
+
         try:
             return self.modules.i18n.get_translations(
                 self.registry.settings.i18n.domain,